ai_services/body_scanner/model.py
import numpy as np
import cv2
from PIL import Image
from io import BytesIO
import random

# Placeholder for a real AI/ML model
# In a real application, this would load a pre-trained model (e.g., TensorFlow, PyTorch)
# and use it to perform human pose estimation and then derive measurements.

def process_image_for_measurements(image_bytes):
    """
    Processes user images to extract body dimensions.
    This is a mock implementation. A real implementation would involve:
    1. Image decoding and preprocessing.
    2. Running a human pose estimation model (e.g., OpenPose, HRNet).
    3. Using detected keypoints and potentially depth information
       (if available from a 3D scanner) to estimate body measurements.
    4. Converting pixel distances to real-world measurements based on estimated height.

    Args:
        image_bytes (bytes): Raw image data (e.g., JPEG, PNG).

    Returns:
        dict: A dictionary of estimated body measurements (height, chest, waist, hips).
              Example: { "height": 170, "chest": 90, "waist": 75, "hips": 95 }
    """
    if not image_bytes:
        raise ValueError("No image bytes provided.")

    try:
        # Attempt to open image with PIL to validate
        Image.open(BytesIO(image_bytes)).verify()
    except Exception as e:
        raise ValueError(f"Invalid image data: {e}")

    # --- MOCK LOGIC START ---
    # For demonstration, we'll return random but reasonable measurements.
    # In a real scenario, this would be derived from the AI model.

    # Simulate some processing time
    # import time
    # time.sleep(1)

    # Generate plausible, slightly varied measurements
    base_height = random.uniform(160, 190) # cm
    base_chest_ratio = random.uniform(0.5, 0.6) # ratio to height
    base_waist_ratio = random.uniform(0.4, 0.5)
    base_hips_ratio = random.uniform(0.55, 0.65)

    height = round(base_height, 1)
    chest = round(base_height * base_chest_ratio * random.uniform(0.9, 1.1), 1)
    waist = round(base_height * base_waist_ratio * random.uniform(0.9, 1.1), 1)
    hips = round(base_height * base_hips_ratio * random.uniform(0.9, 1.1), 1)

    # Ensure measurements are somewhat proportional (e.g., chest > waist)
    if chest < waist:
        chest, waist = waist, chest # Swap if chest is smaller than waist (unlikely for most people)
    if hips < waist:
        hips, waist = waist, hips # Swap if hips is smaller than waist

    measurements = {
        "height": height,
        "chest": chest,
        "waist": waist,
        "hips": hips,
        "unit": "cm" # Specify unit for clarity
    }
    # --- MOCK LOGIC END ---

    print(f"Mock body scan measurements generated: {measurements}")
    return measurements

if __name__ == '__main__':
    # Example usage (for local testing of the model logic)
    # Create a dummy image byte stream
    try:
        dummy_image = Image.new('RGB', (60, 30), color = 'red')
        buf = BytesIO()
        dummy_image.save(buf, format='JPEG')
        dummy_image_bytes = buf.getvalue()

        print("Testing process_image_for_measurements with dummy image:")
        result_measurements = process_image_for_measurements(dummy_image_bytes)
        print(f"Result: {result_measurements}")

        # Test with invalid data
        print("\nTesting with invalid image data:")
        try:
            process_image_for_measurements(b'not an image')
        except ValueError as e:
            print(f"Caught expected error: {e}")

    except Exception as e:
        print(f"An error occurred during local model test: {e}")
